feat: removed unused code and upgraded component
